+++ llvm/lib/Target/X86/X86TargetMachine.cpp
@@ -109,6 +109,7 @@ extern "C" LLVM_C_ABI void LLVMInitializeX86Target() {
   initializeX86WinEHUnwindV2LegacyPass(PR);
   initializeX86PreLegalizerCombinerLegacyPass(PR);
   initializeX86PostLegalizerCombinerLegacyPass(PR);
+  initializeX86WinEHUnwindV3Pass(PR);
 }
 
 static std::unique_ptr<TargetLoweringObjectFile> createTLOF(const Triple &TT) {
@@ -634,8 +635,10 @@ void X86PassConfig::addPreEmitPass2() {
 
   // Analyzes and emits pseudos to support Win x64 Unwind V2. This pass must run
   // after all real instructions have been added to the epilog.
-  if (TT.isOSWindows() && TT.isX86_64())
+  if (TT.isOSWindows() && TT.isX86_64()) {
     addPass(createX86WinEHUnwindV2LegacyPass());
+    addPass(createX86WinEHUnwindV3Pass());
+  }
 }
 
 bool X86PassConfig::addPostFastRegAllocRewrite() {